erro do menor e maior

+++ code/logicoperations.c
@@ -167,9 +167,9 @@ void smallerthan (STACK *s) {
     DATA p2 = pop(s);
     DATA data;
     if (what_type(p1)==LONG && what_type(p2)==LONG) {long val = p2.elems.LONG < p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
-    else if (what_type(p1)==DOUBLE && what_type(p2)==LONG) {long val = p2.elems.LONG < p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
-    else if (what_type(p1)==LONG && what_type(p2)==DOUBLE) {long val = p2.elems.LONG < p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
-    else if (what_type(p1)==DOUBLE && what_type(p2)==DOUBLE) {long val = p2.elems.LONG < p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
+    else if (what_type(p1)==DOUBLE && what_type(p2)==LONG) {long val = p2.elems.LONG < p1.elems.DOUBLE; make_datas(data,LONG,val); push(s,data);}
+    else if (what_type(p1)==LONG && what_type(p2)==DOUBLE) {long val = p2.elems.DOUBLE < p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
+    else if (what_type(p1)==DOUBLE && what_type(p2)==DOUBLE) {long val = p2.elems.DOUBLE < p1.elems.DOUBLE; make_datas(data,LONG,val); push(s,data);}
     else if (what_type(p1)==LONG && what_type(p2)==CHAR) {long val = p2.elems.CHAR < p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
     else if (what_type(p1)==CHAR && what_type(p2)==LONG) {long val = p2.elems.LONG < p1.elems.CHAR; make_datas(data,LONG,val); push(s,data);}
     else if (what_type(p1)==CHAR && what_type(p2)==CHAR) {long val = p2.elems.CHAR < p1.elems.CHAR; make_datas(data,LONG,val); push(s,data);}
@@ -185,9 +185,9 @@ void biggerthan (STACK *s) {
     DATA p2 = pop(s);
     DATA data;
     if (what_type(p1)==LONG && what_type(p2)==LONG) {long val = p2.elems.LONG > p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
-    else if (what_type(p1)==LONG && what_type(p2)==DOUBLE) {long val = p2.elems.LONG > p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
-    else if (what_type(p1)==DOUBLE && what_type(p2)==LONG) {long val = p2.elems.LONG > p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
-    else if (what_type(p1)==DOUBLE && what_type(p2)==DOUBLE) {long val = p2.elems.LONG > p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
+    else if (what_type(p1)==LONG && what_type(p2)==DOUBLE) {long val = p2.elems.DOUBLE > p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
+    else if (what_type(p1)==DOUBLE && what_type(p2)==LONG) {long val = p2.elems.LONG > p1.elems.DOUBLE; make_datas(data,LONG,val); push(s,data);}
+    else if (what_type(p1)==DOUBLE && what_type(p2)==DOUBLE) {long val = p2.elems.DOUBLE > p1.elems.DOUBLE; make_datas(data,LONG,val); push(s,data);}
     else if (what_type(p1)==LONG && what_type(p2)==CHAR) {long val = p2.elems.CHAR > p1.elems.LONG; make_datas(data,LONG,val); push(s,data);}
     else if (what_type(p1)==CHAR && what_type(p2)==LONG) {long val = p2.elems.LONG > p1.elems.CHAR; make_datas(data,LONG,val); push(s,data);}
     else if (what_type(p1)==CHAR && what_type(p2)==CHAR) {long val = p2.elems.CHAR > p1.elems.CHAR; make_datas(data,LONG,val); push(s,data);}
